+++ src/datagen/java/blusunrize/immersiveengineering/data/IEBlockTags.java
@@ -45,6 +45,10 @@ class IEBlockTags extends BlockTagsProvider
 				.addItemEntry(WoodenDecoration.treatedFence.get());
 		getOrCreateBuilder(BlockTags.WOODEN_FENCES)
 				.addItemEntry(WoodenDecoration.treatedFence.get());
+		getOrCreateBuilder(IETags.fencesSteel)
+				.addItemEntry(MetalDecoration.steelFence.get());
+		getOrCreateBuilder(IETags.fencesAlu)
+				.addItemEntry(MetalDecoration.aluFence.get());
 		getOrCreateBuilder(IETags.clayBlock)
 				.addItemEntry(Blocks.CLAY);
 		getOrCreateBuilder(IETags.glowstoneBlock)