+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/ChargingStationTileEntity.java
@@ -12,6 +12,7 @@ import blusunrize.immersiveengineering.api.IEEnums.IOSideConfig;
 import blusunrize.immersiveengineering.api.IEProperties;
 import blusunrize.immersiveengineering.api.energy.immersiveflux.FluxStorage;
 import blusunrize.immersiveengineering.api.energy.immersiveflux.FluxStorageAdvanced;
+import blusunrize.immersiveengineering.client.fx.CustomParticleManager;
 import blusunrize.immersiveengineering.common.IETileTypes;
 import blusunrize.immersiveengineering.common.blocks.IEBaseTileEntity;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IBlockBounds;
@@ -52,6 +53,7 @@ public class ChargingStationTileEntity extends IEBaseTileEntity implements ITick
 {
 	public FluxStorageAdvanced energyStorage = new FluxStorageAdvanced(32000);
 	public NonNullList<ItemStack> inventory = NonNullList.withSize(1, ItemStack.EMPTY);
+	public final CustomParticleManager particles = new CustomParticleManager();
 	private boolean charging = true;
 	public int comparatorOutput = 0;
 
@@ -63,6 +65,8 @@ public class ChargingStationTileEntity extends IEBaseTileEntity implements ITick
 	@Override
 	public void tick()
 	{
+		if(world.isRemote)
+			particles.clientTick();
 		if(EnergyHelper.isFluxReceiver(inventory.get(0)))
 		{
 			if(world.isRemote&&charging)
@@ -78,10 +82,12 @@ public class ChargingStationTileEntity extends IEBaseTileEntity implements ITick
 					if(charge >= 1||(time%12 >= i*4&&time%12 <= i*4+2))
 					{
 						int shift = i-1;
-						double x = getPos().getX()+.5+(getFacing()==Direction.WEST?-.46875: getFacing()==Direction.EAST?.46875: getFacing()==Direction.NORTH?(-.1875*shift): (.1875*shift));
-						double y = getPos().getY()+.25;
-						double z = getPos().getZ()+.5+(getFacing()==Direction.NORTH?-.46875: getFacing()==Direction.SOUTH?.46875: getFacing()==Direction.EAST?(-.1875*shift): (.1875*shift));
-						world.addParticle(new RedstoneParticleData(1-charge, charge, 0, .5f), x, y, z, .25, .25, .25);
+						double x = .5+(getFacing()==Direction.WEST?-.46875: getFacing()==Direction.EAST?.46875: getFacing()==Direction.NORTH?(-.1875*shift): (.1875*shift));
+						double y = .25;
+						double z = .5+(getFacing()==Direction.NORTH?-.46875: getFacing()==Direction.SOUTH?.46875: getFacing()==Direction.EAST?(-.1875*shift): (.1875*shift));
+						particles.add(
+								new RedstoneParticleData(1-charge, charge, 0, .5f), x, y, z, .25, .25, .25, -1
+						);
 					}
 				}
 			}